+++ packages/deployments/src/application/adapter/DeploymentsAdapter.ts
@@ -103,18 +103,18 @@ import { IDistributedPackageRepository } from '../../domain/repositories/IDistri
 import { PublishArtifactsJobFactory } from '../../infra/jobs/PublishArtifactsJobFactory';
 import { DeploymentsServices } from '../services/DeploymentsServices';
 import { TargetResolutionService } from '../services/TargetResolutionService';
-import { AddArtefactsToPackageUsecase } from '../useCases/addArtefactsToPackage/addArtefactsToPackage.usecase';
+import { AddArtefactsToPackageUsecase } from '../useCases/addArtefactsToPackage/AddArtefactsToPackageUsecase';
 import { AddTargetUseCase } from '../useCases/AddTargetUseCase';
-import { CreatePackageUsecase } from '../useCases/createPackage/createPackage.usecase';
-import { UpdatePackageUsecase } from '../useCases/updatePackage/updatePackage.usecase';
+import { CreatePackageUsecase } from '../useCases/createPackage/CreatePackageUsecase';
+import { UpdatePackageUsecase } from '../useCases/updatePackage/UpdatePackageUsecase';
 import { CreateRenderModeConfigurationUseCase } from '../useCases/CreateRenderModeConfigurationUseCase';
-import { DeletePackagesBatchUsecase } from '../useCases/deletePackage/deletePackagesBatch.usecase';
+import { DeletePackagesBatchUsecase } from '../useCases/deletePackage/DeletePackagesBatchUsecase';
 import { DeleteTargetUseCase } from '../useCases/DeleteTargetUseCase';
 import { DeployDefaultSkillsUseCase } from '../useCases/DeployDefaultSkillsUseCase';
 import { DownloadDefaultSkillsZipForAgentUseCase } from '../useCases/DownloadDefaultSkillsZipForAgentUseCase';
 import { DownloadSkillZipForAgentUseCase } from '../useCases/DownloadSkillZipForAgentUseCase';
 import { FindActiveStandardVersionsByTargetUseCase } from '../useCases/FindActiveStandardVersionsByTargetUseCase';
-import { GetPackageByIdUsecase } from '../useCases/getPackageById/getPackageById.usecase';
+import { GetPackageByIdUsecase } from '../useCases/getPackageById/GetPackageByIdUsecase';
 import { GetRenderModeConfigurationUseCase } from '../useCases/GetRenderModeConfigurationUseCase';
 import { GetTargetByIdUseCase } from '../useCases/GetTargetByIdUseCase';
 import { GetTargetsByGitRepoUseCase } from '../useCases/GetTargetsByGitRepoUseCase';
@@ -125,12 +125,12 @@ import { ListActiveDistributedPackagesBySpaceUseCase } from '../useCases/ListAct
 import { ListDistributionsByRecipeUseCase } from '../useCases/ListDistributionsByRecipeUseCase';
 import { ListDistributionsByStandardUseCase } from '../useCases/ListDistributionsByStandardUseCase';
 import { ListDistributionsBySkillUseCase } from '../useCases/ListDistributionsBySkillUseCase';
-import { ListPackagesUsecase } from '../useCases/listPackages/listPackages.usecase';
-import { NotifyArtefactsDistributionUseCase } from '../useCases/notifyArtefactsDistribution/notifyArtefactsDistribution.usecase';
-import { NotifyDistributionUseCase } from '../useCases/notifyDistribution/notifyDistribution.usecase';
+import { ListPackagesUsecase } from '../useCases/listPackages/ListPackagesUsecase';
+import { NotifyArtefactsDistributionUseCase } from '../useCases/notifyArtefactsDistribution/NotifyArtefactsDistributionUseCase';
+import { NotifyDistributionUseCase } from '../useCases/notifyDistribution/NotifyDistributionUseCase';
 import { RemovePackageFromTargetsUseCase } from '../useCases/RemovePackageFromTargetsUseCase';
-import { ListPackagesBySpaceUsecase } from '../useCases/listPackagesBySpace/listPackagesBySpace.usecase';
-import { GetPackageSummaryUsecase } from '../useCases/getPackageSummary/getPackageSummary.usecase';
+import { ListPackagesBySpaceUsecase } from '../useCases/listPackagesBySpace/ListPackagesBySpaceUsecase';
+import { GetPackageSummaryUsecase } from '../useCases/getPackageSummary/GetPackageSummaryUsecase';
 import { PublishArtifactsUseCase } from '../useCases/PublishArtifactsUseCase';
 import { PublishPackagesUseCase } from '../useCases/PublishPackagesUseCase';
 import { GetContentByVersionsUseCase } from '../useCases/GetContentByVersionsUseCase';
