+++ packages/deployments/src/application/useCases/notifyArtefactsDistribution/NotifyArtefactsDistributionUseCase.spec.ts
@@ -1,4 +1,4 @@
-import { NotifyArtefactsDistributionUseCase } from './notifyArtefactsDistribution.usecase';
+import { NotifyArtefactsDistributionUseCase } from './NotifyArtefactsDistributionUseCase';
 import {
   createOrganizationId,
   createPackageId,
