+++ packages/deployments/src/application/useCases/updatePackage/UpdatePackageUsecase.spec.ts
@@ -1,4 +1,4 @@
-import { UpdatePackageUsecase } from './updatePackage.usecase';
+import { UpdatePackageUsecase } from './UpdatePackageUsecase';
 import {
   createUserId,
   createOrganizationId,
